+++ database/migrations/2025_06_13_064737_create_pemohons_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('nomor_hp');
             $table->enum('status', ['proses', 'selesai'])->default('proses');
             $table->string('nama_proses')->default('verifikasi berkas');
-            $table->integer('izin_id');
+            $table->unsignedBigInteger('izin_id');
             $table->timestamps();
             $table->string('no_permohonan');
             $table->foreign('izin_id')->references('id')->on('izins')->onDelete('cascade');
